Factor out a separate 'onboarding' django app.
